Created ecvalidatingfiles module + finished writing filterPredictions()
